+++ src/utils.js
@@ -6,4 +6,8 @@ const randomInteger = (min, max) => {
     return Math.floor(Math.random() * (max - min + 1) + min)
 }
 
-module.exports = {randomDecimal, randomInteger}
+const sleep = async (interval) => {
+    await new Promise(resolve => setTimeout(resolve, interval))
+}
+
+module.exports = {randomDecimal, randomInteger, sleep}
